backend/requirements.txt
asgiref==3.8.1
Brotli==1.1.0
certifi==2024.8.30
click==8.1.7
cloudinary==1.41.0
colorama==0.4.6
Django==5.1.1
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
h11==0.14.0
packaging==24.1
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-dotenv==1.0.1
pytz==2024.2
six==1.16.0
sqlparse==0.5.1
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.31.1
whitenoise==6.7.0
